+++ trilab_invoice/models/__init__.py
@@ -8,3 +8,4 @@ from . import sale_order
 from . import sale_order_line
 
 from . import analytic_account
+from . import account_analytic_default